requirements.txt
numpy
matplotlib
pandas
scipy
requests
python-dotenv